+++ src/main/java/trainers/SeniorTrainer.java
@@ -29,7 +29,7 @@ public class SeniorTrainer implements Trainer
    }
 
    @Override
-   public double deliverExperience( int difficulty )
+   public int deliverExperience( int difficulty )
    {
       return difficulty;
    }
